+++ example/components/simple.js
@@ -2,14 +2,11 @@ import React, { Component } from 'react'
 import { Map, TileLayer, Marker, Popup } from '../../src'
 
 export default class SimpleExample extends Component {
-  constructor () {
-    super()
-    this.state = {
-      lat: 51.505,
-      lng: -0.09,
-      zoom: 13,
-    }
-  }
+  state = {
+    lat: 51.505,
+    lng: -0.09,
+    zoom: 13,
+  };
 
   render () {
     const position = [this.state.lat, this.state.lng]